backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
torch>=2.0.0
transformers>=4.35.0
peft>=0.6.0
pillow>=10.0.0
numpy>=1.24.0
pydantic>=2.0.0
python-multipart>=0.0.6
accelerate>=0.24.0
safetensors>=0.4.0
timm>=0.9.0
requests>=2.31.0